Add Swagger documentation to Thoth server
The springdoc-openapi webmvc starter has been included in the build.gradle file to enable Swagger documentation. A new configuration file, OpenAPISecurityConfiguration.java, was added to set up the OpenAPI definition. This setup includes a brief overview of the application, contact information for the developer, a license, server information and JWT-based security scheme configuration. The inclusion of Swagger will generate a live and interactive API documentation, thereby aiding developers in easily testing and understanding the endpoints.