Update RabbitMQ listener settings and logger level
The changes introduced set the spring.rabbitmq.listener.simple.concurrency to 4 and the logging level for com.thoth to DEBUG in application.properties. This ensures efficient handling of concurrent messages and allows detailed logging for the specified class. Codes in the ClientService.java file were also updated to reflect the new logging level.