Add usage details and permissions check for data sources and templates
This commit introduced usage details view on datasource and template detail pages, helping users to better understand their usage. Also, it added a check for permission to decide if a 'View' or 'Edit' button should be displayed for a particular renderer. It improved the interface usability and enhanced data security by controlling resource access according to permissions.