Add locked icon and improve deletions and layout in pages
This commit adds a locked icon to the datasource and template details pages to signify inaccessible items. It also wraps deletion operations within loading wrappers to improve user feedback during deletions. An Api Docs link is added to the main navigation for easy access, and the typo in Api Keys is fixed. Error handling in the screen message service is also enhanced.