Update DataSourceController and related UI elements
Added handling of `MalformedURLException` and `URISyntaxException` in `DataSourceController` and updated interface elements in `rest-datasource-parameters.component.html`. In the java file, new exceptions were introduced to `checkRestParameters` method. In the HTML file, the UI elements for Query Parameters, Headers, and Body are now wrapped in ion-cards. Also, minor changes in the display formatting have been implemented for better readability.