Update button functions and enhance editor close handling
The button options of the template list page UI have been updated for better user interaction. The editor functionality has been improved with the creation of a 'close' function that better manages the opening and closing of templates. This includes the unsubscription and removal of event listeners when the editor is closed, addressing previous issues with reusability.